moved routing module to main path
