+++ src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
-import { HomepageComponent } from '../homepage/homepage.component';
-import { FormComponent } from '../form/form.component';
-import { RequestStatusComponent } from '../request-status/request-status.component';
+import { HomepageComponent } from './homepage/homepage.component';
+import { FormComponent } from './form/form.component';
+import { RequestStatusComponent } from './request-status/request-status.component';
 
 const routes: Routes = [
   {
