+++ src/app/app.module.ts
@@ -4,7 +4,6 @@ import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientModule } from '@angular/common/http';
 
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatButtonModule } from '@angular/material/button';
@@ -17,7 +16,7 @@ import { Step1Component } from './form/step1/step1.component';
 import { Step2Component } from './form/step2/step2.component';
 import { Step3Component } from './form/step3/step3.component';
 import { HomepageComponent } from './homepage/homepage.component';
-import { AppRoutingModule } from './app-routing/app-routing.module';
+import { AppRoutingModule } from './app-routing.module';
 import { FormComponent } from './form/form.component';
 import { Step4Component } from './form/step4/step4.component';
 import { InformationComponent } from './form/information/information.component';
